third_party/WebKit/LayoutTests/http/tests/security/cookies/websocket/third-party-cookie-accepted.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/websocket/resources/get-request-header.js"></script>
<script src="/websocket/resources/cookie-test-util.js"></script>
<script>
if (window.testRunner) {
  // Don't block third-party cookies.
  testRunner.setBlockThirdPartyCookies(false);
}

const origin = WS_REMOTE_ORIGIN;

promise_test(t => {
  return setCookie(origin).then(
    () => {
      return connectAndGetRequestHeadersFrom(origin);
    }
  ).then(
    headers => {
      const cookie = headers['cookie'];
      assert_not_equals(cookie, undefined,
                        'headers should contain a cookie header.');
      // Same-site cookies are not included.
      assert_equals(normalizeCookie(cookie), 'ws-path-root=1; ws=1');
      return clear(origin);
    }
  );
}, "Test that third-party cookies are accepted for WebSockets when blocking " +
    "is disabled.");
</script>
